requirements.txt
streamlit_lottie==0.0.3
streamlit~=1.23.1
requests==2.31.0
Pillow~=9.5.0
protobuf==4.23.2
watchdog==3.0.0
streamlit_cookies_manager~=0.2.0
cryptography~=41.0.1
python-dotenv==1.0.0
